feat: produce a fuller physical operator for updates
