+++ go/vt/vtgate/planbuilder/abstract/operator.go
@@ -273,7 +273,7 @@ func createOperatorFromUpdate(updStmt *sqlparser.Update, semTable *semantics.Sem
 		assignments[set.Name.Name.String()] = set.Expr
 	}
 
-	return &Update{Table: qt, Assignments: assignments}, nil
+	return &Update{Table: qt, Assignments: assignments, AST: updStmt, TableInfo: tableInfo}, nil
 }
 
 func addColumnEquality(semTable *semantics.SemTable, expr sqlparser.Expr) {
