+++ go/vt/vtgate/planbuilder/testdata/onecase.txt
@@ -1,25 +1,1 @@
 # Add your test case here for debugging and run go test -run=One.
-"update user set col = 5 where id = 4"
-{
-  "QueryType": "UPDATE",
-  "Original": "update user set col = 5 where id = 4",
-  "Instructions": {
-    "OperatorType": "Update",
-    "Variant": "Equal",
-    "Keyspace": {
-      "Name": "user",
-      "Sharded": true
-    },
-    "TargetTabletType": "PRIMARY",
-    "MultiShardAutocommit": false,
-    "Query": "update `user` set col = 5 where id = 4",
-    "Table": "user",
-    "Values": [
-      "INT64(4)"
-    ],
-    "Vindex": "user_index"
-  }
-}
-{
-
-}
